moved recursion helper to utilities
